Use webpack require instead of graphql to import data
TODO: Write a helper function, or perhaps a webpack loader,
so I can just say ' import data from "*./json" '
