+++ src/cms/preview-templates/ImprintPreview.tsx
@@ -1,10 +1,9 @@
 import React, { FunctionComponent } from 'react'
 import { ImprintComponent } from '../../components/Imprint'
-import { getTranslations } from '../../data/queryHelpers'
 import * as R from 'ramda'
 import { Imprint } from '../../data/imprint'
 import { PreviewProps, getJSON } from './helpers'
-import { Localized } from '../../data'
+import { Localized, getTranslations } from '../../data/i18n'
 
 export const ImprintPreview: FunctionComponent<PreviewProps> = ({ entry }) => {
   const imprint: Localized<Imprint> = getJSON(entry)
