+++ static/admin/config.yml
@@ -3,6 +3,7 @@ backend:
   repo: hardforkio/homepage
   branch: master
 
+site_url: "https://hardfork.io"
 media_folder: "static/img"
 public_folder: "/img"
 publish_mode: editorial_workflow
